+++ ocr/src/main/java/com/ing/software/ocr/OcrObjects/TicketSchemes/TicketSchemeIT_PSCC.java
@@ -1,5 +1,8 @@
 package com.ing.software.ocr.OcrObjects.TicketSchemes;
 
+import android.support.annotation.NonNull;
+
+import com.annimon.stream.Stream;
 import com.ing.software.common.Scored;
 
 import java.math.BigDecimal;
@@ -20,7 +23,7 @@ public class TicketSchemeIT_PSCC implements TicketScheme{
     private BigDecimal cash;
     private BigDecimal change;
 
-    public TicketSchemeIT_PSCC(BigDecimal total, List<BigDecimal> aboveTotal, List<BigDecimal> belowTotal) {
+    public TicketSchemeIT_PSCC(BigDecimal total, @NonNull List<BigDecimal> aboveTotal, @NonNull List<BigDecimal> belowTotal) {
         this.total = total;
         if (!aboveTotal.isEmpty())
             this.subtotal = aboveTotal.get(aboveTotal.size()-1);
@@ -35,11 +38,149 @@ public class TicketSchemeIT_PSCC implements TicketScheme{
 
     @Override
     public Scored<BigDecimal> getBestAmount() {
-        return null;
+        return temporaryBestAmount();
     }
 
     @Override
     public String toString() {
         return tag;
+    }
+
+    /*
+    Temporary, copied and modified from old amount comparator
+     */
+    private Scored<BigDecimal> temporaryBestAmount() {
+        int FOUR_VALUES = 100;
+        int THREE_VALUES = 65;
+        int THREE_VALUES_AMOUNT = 80;
+        int TWO_VALUES_AMOUNT = 55;
+        int TWO_VALUES = 30;
+        int NO_MATCH = 1;
+        if (total != null) {
+            if (products != null && cash != null && subtotal != null && change != null) {
+                BigDecimal productsSum = Stream.of(products)
+                        .reduce(BigDecimal.ZERO, BigDecimal::add);
+                BigDecimal normCash = cash.subtract(change);
+                boolean cashPrices = normCash.compareTo(productsSum) == 0;
+                boolean cashAmount = normCash.compareTo(total) == 0;
+                boolean pricesAmount = productsSum.compareTo(total) == 0;
+                boolean subtotalAmount = subtotal.compareTo(total) == 0;
+                boolean subtotalCash = subtotal.compareTo(normCash) == 0;
+                boolean subtotalPrices = subtotal.compareTo(productsSum) == 0;
+                if (cashPrices && pricesAmount) {
+                    if (subtotalAmount) {
+                        return new Scored<>(FOUR_VALUES, total);
+                    } else {
+                        return new Scored<>(THREE_VALUES_AMOUNT, total);
+                    }
+                } else if (cashAmount && subtotalAmount) {
+                    return new Scored<>(THREE_VALUES_AMOUNT, total);
+                } else if (pricesAmount && subtotalAmount) {
+                    return new Scored<>(THREE_VALUES_AMOUNT, total);
+                } else if (cashPrices && subtotalPrices) {
+                    return new Scored<>(THREE_VALUES, subtotal);
+                } else if (cashAmount || pricesAmount || subtotalAmount) {
+                    return new Scored<>(TWO_VALUES_AMOUNT, total);
+                } else if (cashPrices || subtotalCash) {
+                    return new Scored<>(TWO_VALUES, normCash);
+                } else if (subtotalPrices) {
+                    return new Scored<>(TWO_VALUES, subtotal);
+                } else {
+                    return new Scored<>(NO_MATCH, total);
+                }
+            } else if (products != null && cash != null && change != null) {
+                BigDecimal productsSum = Stream.of(products)
+                        .reduce(BigDecimal.ZERO, BigDecimal::add);
+                BigDecimal normCash = cash.subtract(change);
+                boolean cashPrices = normCash.compareTo(productsSum) == 0;
+                boolean cashAmount = normCash.compareTo(total) == 0;
+                boolean pricesAmount = productsSum.compareTo(total) == 0;
+                if (cashPrices || pricesAmount) {
+                    if (cashAmount) {
+                        return new Scored<>(THREE_VALUES, total);
+                    } else if (pricesAmount) {
+                        return new Scored<>(TWO_VALUES_AMOUNT, total);
+                    } else {
+                        return new Scored<>(TWO_VALUES, normCash);
+                    }
+                } else {
+                    return new Scored<>(NO_MATCH, total);
+                }
+            } else if (products != null && subtotal != null) {
+                BigDecimal productsSum = Stream.of(products)
+                        .reduce(BigDecimal.ZERO, BigDecimal::add);
+                boolean pricesAmount = productsSum.compareTo(total) == 0;
+                boolean subtotalAmount = subtotal.compareTo(total) == 0;
+                boolean subtotalPrices = subtotal.compareTo(productsSum) == 0;
+                if (subtotalPrices || pricesAmount) {
+                    if (subtotalAmount) {
+                        return new Scored<>(THREE_VALUES, total);
+                    } else if (pricesAmount) {
+                        return new Scored<>(TWO_VALUES_AMOUNT, total);
+                    } else {
+                        return new Scored<>(TWO_VALUES, subtotal);
+                    }
+                } else {
+                    return new Scored<>(NO_MATCH, total);
+                }
+            } else if (cash != null && subtotal != null && change != null) {
+                BigDecimal normCash = cash.subtract(change);
+                boolean cashAmount = normCash.compareTo(total) == 0;
+                boolean subtotalAmount = subtotal.compareTo(total) == 0;
+                boolean subtotalCash = subtotal.compareTo(normCash) == 0;
+                if (subtotalCash || cashAmount) {
+                    if (subtotalAmount) {
+                        return new Scored<>(THREE_VALUES, total);
+                    } else if (cashAmount) {
+                        return new Scored<>(TWO_VALUES_AMOUNT, total);
+                    } else {
+                        return new Scored<>(TWO_VALUES, subtotal);
+                    }
+                } else {
+                    return new Scored<>(NO_MATCH, total);
+                }
+            } else {
+                return new Scored<>(NO_MATCH, total);
+            }
+        } else {
+            //no total
+            if (products != null && cash != null && subtotal != null && change != null) {
+                BigDecimal productsSum = Stream.of(products)
+                        .reduce(BigDecimal.ZERO, BigDecimal::add);
+                BigDecimal normCash = cash.subtract(change);
+                boolean cashPrices = normCash.compareTo(productsSum) == 0;
+                boolean subtotalCash = subtotal.compareTo(normCash) == 0;
+                boolean subtotalPrices = subtotal.compareTo(productsSum) == 0;
+                if (cashPrices && subtotalCash) {
+                    return new Scored<>(THREE_VALUES, normCash);
+                } else if (cashPrices || subtotalCash) {
+                    return new Scored<>(TWO_VALUES, normCash);
+                } else if (subtotalPrices) {
+                    return new Scored<>(TWO_VALUES, subtotal);
+                }
+            } else if (products != null && cash != null && change != null) {
+                BigDecimal productsSum = Stream.of(products)
+                        .reduce(BigDecimal.ZERO, BigDecimal::add);
+                BigDecimal normCash = cash.subtract(change);
+                boolean cashPrices = normCash.compareTo(productsSum) == 0;
+                if (cashPrices) {
+                    return new Scored<>(TWO_VALUES, normCash);
+                }
+            } else if (products != null && subtotal != null) {
+                BigDecimal productsSum = Stream.of(products)
+                        .reduce(BigDecimal.ZERO, BigDecimal::add);
+                boolean subtotalPrices = subtotal.compareTo(productsSum) == 0;
+                if (subtotalPrices) {
+                    return new Scored<>(TWO_VALUES, subtotal);
+                }
+            } else if (cash != null && subtotal != null && change != null) {
+                BigDecimal normCash = cash.subtract(change);
+                boolean subtotalCash = subtotal.compareTo(normCash) == 0;
+                if (subtotalCash) {
+                    return new Scored<>(TWO_VALUES, subtotal);
+                }
+            }
+            return null;
+        }
     }
 }