+++ ocr/src/main/java/com/ing/software/ocr/OcrVars.java
@@ -25,9 +25,6 @@ public class OcrVars {
     public static final String CONCLUSION_TAG = "conclusion"; //tag for rawtext on bottom of the receipt
     public static final double NUMBER_MIN_VALUE = 0.4; //Max allowed value to accept a string as a valid number. See OcrUtils.isPossiblePriceNumber()
     public static final double NUMBER_MIN_VALUE_ALTERNATIVE = 0.1; //Max allowed value to accept a string as a valid number. See OcrUtils.isPossiblePriceNumber(). Used in alternative search
-    public static final int HEIGHT_CENTER_DIFF_MULTIPLIER = 50; //Multiplier used while analyzing difference in alignment between the center of two rects in dataAnalyzer
-    public static final int HEIGHT_LIST_MULTIPLIER = 80; //Multiplier used while analyzing difference between average height of rects and a specific rect. Used in ProbGrid.getRectHeightScore()
-    public static final int HEIGHT_SOURCE_DIFF_MULTIPLIER = 50; //Multiplier used while analyzing difference in height between source and target rect (total with it's price)
     public static final float RECT_HEIGHT_EXTENDER = 1.5f; //Extend height of chosen rect
     public static final int NUMBER_MAX_LENGTH = 8; //Max number of digits allowed for numbers
     public static final double MIN_DIGITS_NUMBER = 2./3.;