+++ ocr/src/main/java/com/ing/software/ocr/OperativeObjects/AmountComparator.java
@@ -3,6 +3,9 @@ package com.ing.software.ocr.OperativeObjects;
 import android.util.Pair;
 
 import com.ing.software.common.Scored;
+import com.ing.software.ocr.DataAnalyzer;
+import com.ing.software.ocr.OcrManager;
+import com.ing.software.ocr.OcrObjects.TempText;
 import com.ing.software.ocr.OcrObjects.TicketSchemes.*;
 
 import java.math.BigDecimal;
@@ -11,6 +14,9 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import com.annimon.stream.Stream;
+import com.ing.software.ocr.ScoreFunc;
+
+import static com.ing.software.ocr.OcrVars.NUMBER_MIN_VALUE;
 
 /**
  * Class to analyze amount and compare it with a predefined ticket.
@@ -18,7 +24,6 @@ import com.annimon.stream.Stream;
 
 public class AmountComparator {
 
-    private BigDecimal amount;
     private List<Scored<Pair<BigDecimal, TicketScheme>>> bestAmounts = new ArrayList<>();
     private List<TicketScheme> acceptedSchemes = new ArrayList<>();
 
@@ -26,11 +31,9 @@ public class AmountComparator {
      * @author Michelon
      * Constructor, check amount against a defined scheme
      * @date 27-1-18
-     * @param scheme
-     * @param amount BigDecimal containing decoded amount.
+     * @param scheme scheme to use to check amount
      */
-    public AmountComparator(TicketScheme scheme, BigDecimal amount) {
-        this.amount = amount;
+    public AmountComparator(TicketScheme scheme) {
         acceptedSchemes.add(scheme);
     }
 
@@ -38,25 +41,27 @@ public class AmountComparator {
      * @author Michelon
      * Constructor, check amount against a list of schemes
      * @date 27-1-18
-     * @param schemes
-     * @param amount BigDecimal containing decoded amount.
+     * @param schemes list of schemes to use to check amount
      */
-    public AmountComparator(List<TicketScheme> schemes, BigDecimal amount) {
-        this.amount = amount;
+    public AmountComparator(List<TicketScheme> schemes) {
         acceptedSchemes.addAll(schemes);
     }
 
     /**
      * Constructor, check amount against all schemes
-     * @param amount
+     * @param amount BigDecimal containing decoded amount.
      */
-    public AmountComparator(BigDecimal amount) {
-        this.amount = amount;
-        acceptedSchemes = getAllSchemes(amount, getAboveTotalPrices(), getBelowTotalPrices());
+    public AmountComparator(BigDecimal amount, TempText amountText) {
+        acceptedSchemes = getAllSchemes(amount, getAboveTotalPrices(amountText), getBelowTotalPrices(amountText));
     }
 
+    /**
+     * Retrieves amount with highest score from ticket schemes
+     * @return amount with highest score, it's score and the ticket used. Null if no valid amount was found.
+     */
     public Scored<Pair<BigDecimal, TicketScheme>> getBestAmount() {
         bestAmounts = Stream.of(acceptedSchemes)
+                .filter(scheme -> scheme.getBestAmount() != null)
                 .map(scheme -> new Scored<>(scheme.getBestAmount().getScore(), new Pair<>(scheme.getBestAmount().obj(), scheme)))
                 .toList();
         if (!bestAmounts.isEmpty()) {
@@ -66,14 +71,39 @@ public class AmountComparator {
             return null;
     }
 
-    private List<BigDecimal> getAboveTotalPrices() {
-        return null;
+    /**
+     * Analyze the list of products from OcrSchemer and convert accepted values as bigdecimal
+     * @return list of bigdecimal of numbers above total
+     */
+    public List<BigDecimal> getAboveTotalPrices(TempText amountText) {
+        return Stream.of(OcrManager.mainImage.getPricesTexts())
+                .filter(price -> price.box().centerY() < amountText.box().centerY())
+                .filter(price -> ScoreFunc.isPossiblePriceNumber(price.textNoSpaces(), price.numNoSpaces()) < NUMBER_MIN_VALUE)
+                .map(price -> DataAnalyzer.analyzeAmount(price.numNoSpaces()))
+                .withoutNulls()
+                .toList();
     }
 
-    private List<BigDecimal> getBelowTotalPrices() {
-        return null;
+    /**
+     * Analyze the list of products from OcrSchemer and convert accepted values as bigdecimal
+     * @return list of bigdecimal of numbers below total
+     */
+    public List<BigDecimal> getBelowTotalPrices(TempText amountText) {
+        return Stream.of(OcrManager.mainImage.getPricesTexts())
+                .filter(price -> price.box().centerY() > amountText.box().centerY())
+                .filter(price -> ScoreFunc.isPossiblePriceNumber(price.textNoSpaces(), price.numNoSpaces()) < NUMBER_MIN_VALUE)
+                .map(price -> DataAnalyzer.analyzeAmount(price.numNoSpaces()))
+                .withoutNulls()
+                .toList();
     }
 
+    /**
+     * Initialize a list of all TicketSchemes
+     * @param total bigdecimal with total value
+     * @param aboveTotal list of prices above total
+     * @param belowTotal list of prices below total
+     * @return list of all ticketSchemes
+     */
     private static List<TicketScheme> getAllSchemes(BigDecimal total, List<BigDecimal> aboveTotal, List<BigDecimal> belowTotal) {
         TicketScheme[] schemes = {
                 new TicketSchemeIT_PC(total, aboveTotal, belowTotal),